move to models folder
